# src/main/java/com/obify/rms/entity/OrganizationEntity.java
package com.obify.rms.entity;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Document(collection = "organization")
public class OrganizationEntity {

    @Id
    private String id;
    private String orgName;
}