# src/main/resources/json/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/reviews":{"post":{"tags":["review-controller"],"operationId":"addReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}}}}},"/api/v1/organization":{"post":{"tags":["organization-controller"],"operationId":"addOrg","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationEntity"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationEntity"}}}}}}},"/api/v1/reviews/{reviewId}":{"patch":{"tags":["review-controller"],"operationId":"updateStatus","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReviewEntity"}}}}}}},"/api/v1/reviews/{organizationId}":{"get":{"tags":["review-controller"],"operationId":"getReviews","parameters":[{"name":"organizationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewEntity"}}}}}}}}},"components":{"schemas":{"ReviewEntity":{"type":"object","properties":{"id":{"type":"string"},"rating":{"type":"number","format":"float"},"userId":{"type":"string"},"productId":{"type":"string"},"review":{"type":"string"},"organizationId":{"type":"string"},"status":{"type":"string"}}},"OrganizationEntity":{"type":"object","properties":{"id":{"type":"string"},"orgName":{"type":"string"}}},"ReviewDTO":{"type":"object","properties":{"status":{"type":"string"}}}}}}